vision/stages/drawing-primitives/pyproject.toml
[project]
name = "drawing-primitives"
version = "0.1.0"
description = "Per-element visual primitive extraction (color, border, shadow, font)"
requires-python = ">=3.11"
dependencies = [
    "testanyware-common",
    "numpy>=1.26",
    "opencv-python-headless>=4.8",
    "pillow>=10.0",
    "scikit-learn>=1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drawing_primitives"]
